+++ trunk/Crystal/CrystalPhysics/Obstacle.cpp
@@ -4,7 +4,7 @@
 
 using namespace Crystal::Physics;
 
-Obstacle::Obstacle( const int id, const double density, const double gasConstant, const double viscosityCoefficient, ParticleFactory* particleFactory ) :
+Obstacle::Obstacle( const int id, const float density, const float gasConstant, const float viscosityCoefficient, ParticleFactory* particleFactory ) :
 PhysicsObject( id , density, gasConstant, viscosityCoefficient, particleFactory )
 {
 }
