+++ trunk/Crystal/CrystalPhysics/ParticleConditions.h
@@ -15,20 +15,20 @@ class ParticleConditions
 {
 public:
 
-	ParticleConditions(const std::vector<Geom::Vector3d>& points, const double particleLength, const double density) :
+	ParticleConditions(const std::vector<Geom::Vector3d>& points, const float particleLength, const float density) :
 		particleLength( particleLength),
 		 density( density),
 		 innerPoints( points)
 		{
 		}
 
-	double getParticleLength() const { return particleLength; }
+	float getParticleLength() const { return particleLength; }
 
-	double getDensity() const { return density; }
+	float getDensity() const { return density; }
 
-	double getParticleVolume() const { return particleLength * particleLength * particleLength; }
+	float getParticleVolume() const { return particleLength * particleLength * particleLength; }
 
-	double getParticleMass() const { return density * getParticleVolume(); }
+	float getParticleMass() const { return density * getParticleVolume(); }
 
 	std::vector<Geom::Vector3d> getInnerPoints() const { return innerPoints; }
 
@@ -36,8 +36,8 @@ protected:
 	void createInnerPoints(const Geom::Box& box);
 
 private:
-	const double particleLength;
-	const double density;
+	const float particleLength;
+	const float density;
 	const std::vector<Geom::Vector3d> innerPoints;
 };
